Moved most of AdminController to fint-core-consumer-shared
