+++ build.gradle
@@ -48,6 +48,7 @@ dependencies {
         exclude group: 'com.github.springfox.loader'
         exclude group: 'org.springframework.boot'
     }
+    implementation 'no.fintlabs:fint-core-consumer-shared:0-SNAPSHOT'
 
     compileOnly 'org.projectlombok:lombok'
     runtimeOnly 'io.micrometer:micrometer-registry-prometheus'
