+++ src/main/java/no/fintlabs/consumer/fravar/FravarKafkaConsumer.java
@@ -1,7 +1,7 @@
 package no.fintlabs.consumer.fravar;
 
 import no.fint.model.resource.utdanning.vurdering.FravarResource;
-import no.fintlabs.consumer.KafkaConsumer;
+import no.fintlabs.core.consumer.shared.KafkaConsumer;
 import no.fintlabs.kafka.common.ListenerBeanRegistrationService;
 import no.fintlabs.kafka.entity.EntityConsumerFactoryService;
 import no.fintlabs.kafka.entity.topic.EntityTopicService;
