+++ src/test/groovy/no/fintlabs/consumer/AdminControllerSpec.groovy
@@ -2,7 +2,8 @@ package no.fintlabs.consumer
 
 import no.fintlabs.cache.CacheManager
 import no.fintlabs.consumer.admin.AdminController
-import no.fintlabs.consumer.config.ConsumerProps
+import no.fintlabs.core.consumer.shared.ConsumerProps
+import no.fintlabs.core.consumer.shared.ConsumerService
 import spock.lang.Specification
 
 class AdminControllerSpec extends Specification {
